\name{OrdLogReg-package}
\alias{OrdLogReg-package}
\alias{OrdLogReg}
\docType{package}
\title{
Ordinal Logic Regression
}
\description{
Develops a classification model for ordinal responses.  A logic regression trees is constructed for ordinal categories in the data and then all trees are used to predict the class of observations. 
}
\details{
\tabular{ll}{
Package: \tab OrdLogReg\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2013-04-12\cr
License: \tab GPL-2\cr
}

}
\author{
Bethany Wolf

Maintainer: Bethany Wolf <wolfb@musc.edu>
}

\references{Wolf B.J., Slate E.H., Hill E.G. (2014).  Ordinal Logic Regression:  a classifier for discovering logical combinations of binary markers for ordinal outcomes.  Computational Statistics and Data Analysis.}

\keyword{ package }

