\name{OrdLogReg-package}
\alias{OrdLogReg-package}
\alias{OrdLogReg}
\docType{package}
\title{
Ordinal Logic Regression
}
\description{
Develops a classification model for ordinal responses.  A logic regression trees is constructed for ordinal categories in the data and then all trees are used to predict the class of observations. 
}
\details{
\tabular{ll}{
Package: \tab OrdLogReg\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-12\cr
License: \tab GPL-2\cr
}

}
\author{
Bethany Wolf

Maintainer: Bethany Wolf <wolfb@musc.edu>
}


\keyword{ package }

